+++ internal/cmd/jobs/cmd.go
@@ -8,7 +8,6 @@ import (
 
 	"github.com/saucelabs/saucectl/internal/cmd/jobs/job"
 	"github.com/saucelabs/saucectl/internal/credentials"
-	"github.com/saucelabs/saucectl/internal/insights"
 	"github.com/saucelabs/saucectl/internal/region"
 	"github.com/saucelabs/saucectl/internal/saucecloud"
 	"github.com/spf13/cobra"
@@ -38,7 +37,7 @@ func Command(preRun func(cmd *cobra.Command, args []string)) *cobra.Command {
 
 			creds := credentials.Get()
 			url := region.FromString(regio).APIBaseURL()
-			insightsClient := insights.NewInsightsService(url, creds, insightsTimeout)
+			insightsClient := http2.NewInsightsService(url, creds, insightsTimeout)
 			iamClient := http2.UserService{
 				HTTPClient:  &http.Client{Timeout: iamTimeout},
 				URL:         url,
